+++ src/lib.rs
@@ -6,12 +6,15 @@
 #[macro_use] extern crate error_type;
 extern crate rustc_serialize;
 extern crate iron;
+extern crate hyper;
 #[macro_use] extern crate router;
 extern crate urlencoded;
 extern crate redis;
 extern crate spaceapi;
 
 pub use spaceapi as api;
+pub use iron::error::HttpResult;
+pub use hyper::server::Listening;
 
 mod server;
 mod errors;