+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Possible log types:
 ### UNRELEASED
 
 - [changed] Removed datastore module, use Redis directly (#10)
+- [changed] SpaceapiServer.serve() now returns a HttpResult<Listening> (#16)
 - [added] Support status modifiers (#8)
 
 ### v0.1.1 (2015-11-16)